refactor: Make a hooks folder for custom hooks
